# src/main/resources/swagger.yaml
swagger: '2.0'
info:
  description: DataStation API
  version: v0.01
  title: DataStation API
  termsOfService: Terms of service
  contact:
    name: N.A
    url: N.A
    email: N.A
  license:
    name: License of API
    url: API license URL
host: '192.168.210.131:7578'
basePath: /
...
paths:
  /api/auth/signin:
    post:
      tags:
        - auth-controller
      summary: authenticateUser
      operationId: authenticateUserUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: loginRequest
          description: loginRequest
          required: true
          schema:
            $ref: '#/definitions/LoginRequest'
      responses:
        '200':
          description: OK
          schema:
            type: object
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      security:
        - JWT:
            - global
      deprecated: false
  /api/auth/signup:
    post:
      tags:
        - auth-controller
      summary: registerUser
      operationId: registerUserUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: signUpRequest
          description: signUpRequest
          required: true
          schema:
            $ref: '#/definitions/SignUpRequest'
      responses:
        '200':
          description: OK
          schema:
            type: object
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      security:
        - JWT:
            - global
      deprecated: false
...
  '/api/users/{username}':
    get:
      tags:
        - user-controller
      summary: getUserProfile
      operationId: getUserProfileUsingGET
      produces:
        - '*/*'
      parameters:
        - name: username
          in: path
          description: username
          required: true
          type: string
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/UserProfile'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      security:
        - JWT:
            - global
      deprecated: false
...
securityDefinitions:
  JWT:
    type: apiKey
    name: Authorization
    in: header
definitions:
...
  LoginRequest:
    type: object
    properties:
      password:
        type: string
      username_or_email:
        type: string
    title: LoginRequest
...
  SignUpRequest:
    type: object
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
      username:
        type: string
    title: SignUpRequest
...
  Timestamp:
    type: object
    properties:
      date:
        type: integer
        format: int32
      day:
        type: integer
        format: int32
      hours:
        type: integer
        format: int32
      minutes:
        type: integer
        format: int32
      month:
        type: integer
        format: int32
      nanos:
        type: integer
        format: int32
      seconds:
        type: integer
        format: int32
      time:
        type: integer
        format: int64
      timezone_offset:
        type: integer
        format: int32
      year:
        type: integer
        format: int32
    title: Timestamp
...
  UserProfile:
    type: object
    properties:
      id:
        type: integer
        format: int64
      joined_at:
        $ref: '#/definitions/Timestamp'
      name:
        type: string
      username:
        type: string
    title: UserProfile
...